pyproject.toml
[project]
name = "profound-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "sortedcontainers>=2.4.0",
    "uvicorn[standard]>=0.42.0",
]
